[Refactor] Move FragmentExecutor into query orchestration
