+++ be/src/exec/pipeline/fragment_execution_params.h
@@ -1,0 +1,87 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <map>
+#include <vector>
+
+#include "gen_cpp/InternalService_types.h"
+#include "glog/logging.h"
+#include "gutil/macros.h"
+
+namespace starrocks::pipeline {
+
+// For the exec_batch_plan_fragments RPC request, common_request and unique_request are different.
+// - common_request contains the common fields of all the fragment instances.
+// - unique_request contains the unique fields for a specific fragment instance, including:
+//    - backend_num
+//    - pipeline_dop
+//    - params.fragment_instance_id
+//    - params.sender_id
+//    - params.per_node_scan_ranges
+//    - fragment.output_sink (only for MultiCastDataStreamSink and ExportSink)
+// For the exec_plan_fragments request, common_request and unique_request are identical.
+using PerDriverScanRangesMap = std::map<int32_t, std::vector<TScanRangeParams>>;
+
+class UnifiedExecPlanFragmentParams {
+public:
+    UnifiedExecPlanFragmentParams(const TExecPlanFragmentParams& common_request,
+                                  const TExecPlanFragmentParams& unique_request)
+            : _common_request(common_request), _unique_request(unique_request) {
+        DCHECK(unique_request.__isset.backend_num);
+        DCHECK(unique_request.__isset.pipeline_dop);
+        DCHECK(unique_request.__isset.params);
+        DCHECK(unique_request.params.__isset.sender_id);
+    }
+
+    DISALLOW_COPY_AND_MOVE(UnifiedExecPlanFragmentParams);
+
+    const TExecPlanFragmentParams& common() const { return _common_request; }
+    const TExecPlanFragmentParams& unique() const { return _unique_request; }
+
+    int32_t backend_num() const { return _unique_request.backend_num; }
+    int32_t pipeline_dop() const { return _unique_request.__isset.pipeline_dop ? _unique_request.pipeline_dop : 0; }
+    // The maximum number of pipelines that can be instantiated in a colocate group when group execution is enabled.
+    // If a fragment is not using group execution, return pipeline_dop().
+    int32_t group_execution_scan_dop() const {
+        return _unique_request.__isset.group_execution_scan_dop ? _unique_request.group_execution_scan_dop
+                                                                : pipeline_dop();
+    }
+    int32_t pipeline_sink_dop() const {
+        // NOTE: default dop is 1, compatible with old version(before 2.5)
+        return _unique_request.params.__isset.pipeline_sink_dop ? _unique_request.params.pipeline_sink_dop : 1;
+    }
+
+    const std::vector<TExecDebugOption>& debug_actions() const { return _unique_request.params.exec_debug_options; }
+
+    const TUniqueId& fragment_instance_id() const { return _unique_request.params.fragment_instance_id; }
+    int32_t sender_id() const { return _unique_request.params.sender_id; }
+
+    const std::vector<TScanRangeParams>& scan_ranges_of_node(TPlanNodeId node_id) const;
+    const PerDriverScanRangesMap& per_driver_seq_scan_ranges_of_node(TPlanNodeId node_id) const;
+
+    bool isset_output_sink() const {
+        return _common_request.fragment.__isset.output_sink || _unique_request.fragment.__isset.output_sink;
+    }
+    const TDataSink& output_sink() const;
+
+private:
+    static const std::vector<TScanRangeParams> _no_scan_ranges;
+    static const PerDriverScanRangesMap _no_scan_ranges_per_driver_seq;
+    const TExecPlanFragmentParams& _common_request;
+    const TExecPlanFragmentParams& _unique_request;
+};
+
+} // namespace starrocks::pipeline
