--- be/src/exec/pipeline/fragment_executor.h
@@ -1,169 +1,0 @@
-// Copyright 2021-present StarRocks, Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include <map>
-#include <unordered_map>
-#include <vector>
-
-#include "common/global_types.h"
-#include "common/status.h"
-#include "compute_env/workgroup/work_group_fwd.h"
-#include "exec/pipeline/pipeline_fwd.h"
-#include "exec/runtime/group_execution/execution_group_fwd.h"
-#include "gen_cpp/InternalService_types.h"
-#include "gutil/macros.h"
-#include "runtime/exec_env_fwd.h"
-
-namespace starrocks {
-class DataSink;
-class RuntimeProfile;
-class TPlanFragmentExecParams;
-class RuntimeState;
-
-namespace pipeline {
-class FragmentContext;
-class PipelineBuilderContext;
-class QueryContext;
-class QueryContextManager;
-
-// For the exec_batch_plan_fragments RPC request, common_request and unique_request are different.
-// - common_request contains the common fields of all the fragment instances.
-// - unique_request contains the unique fields for a specific fragment instance, including:
-//    - backend_num
-//    - pipeline_dop
-//    - params.fragment_instance_id
-//    - params.sender_id
-//    - params.per_node_scan_ranges
-//    - fragment.output_sink (only for MultiCastDataStreamSink and ExportSink)
-// For the exec_plan_fragments request, common_request and unique_request are identical.
-
-using PerDriverScanRangesMap = std::map<int32_t, std::vector<TScanRangeParams>>;
-class UnifiedExecPlanFragmentParams {
-public:
-    UnifiedExecPlanFragmentParams(const TExecPlanFragmentParams& common_request,
-                                  const TExecPlanFragmentParams& unique_request)
-            : _common_request(common_request), _unique_request(unique_request) {
-        DCHECK(unique_request.__isset.backend_num);
-        DCHECK(unique_request.__isset.pipeline_dop);
-        DCHECK(unique_request.__isset.params);
-        DCHECK(unique_request.params.__isset.sender_id);
-    }
-
-    DISALLOW_COPY_AND_MOVE(UnifiedExecPlanFragmentParams);
-
-    // Access the common fields by this method.
-    const TExecPlanFragmentParams& common() const { return _common_request; }
-    const TExecPlanFragmentParams& unique() const { return _unique_request; }
-
-    // Access the unique fields by the following methods.
-    int32_t backend_num() const { return _unique_request.backend_num; }
-    int32_t pipeline_dop() const { return _unique_request.__isset.pipeline_dop ? _unique_request.pipeline_dop : 0; }
-    // The maximum number of pipelines that can be instantiated in a colocate group when group execution is enabled.
-    // if a fragment is not using group execution. It will return pipeline_dop()
-    int32_t group_execution_scan_dop() const {
-        return _unique_request.__isset.group_execution_scan_dop ? _unique_request.group_execution_scan_dop
-                                                                : pipeline_dop();
-    }
-    int32_t pipeline_sink_dop() const {
-        // NOTE: default dop is 1, compatible with old version(before 2.5)
-        return _unique_request.params.__isset.pipeline_sink_dop ? _unique_request.params.pipeline_sink_dop : 1;
-    }
-
-    const std::vector<TExecDebugOption>& debug_actions() const { return _unique_request.params.exec_debug_options; }
-
-    const TUniqueId& fragment_instance_id() const { return _unique_request.params.fragment_instance_id; }
-    int32_t sender_id() const { return _unique_request.params.sender_id; }
-
-    const std::vector<TScanRangeParams>& scan_ranges_of_node(TPlanNodeId node_id) const;
-    const PerDriverScanRangesMap& per_driver_seq_scan_ranges_of_node(TPlanNodeId node_id) const;
-
-    bool isset_output_sink() const {
-        return _common_request.fragment.__isset.output_sink || _unique_request.fragment.__isset.output_sink;
-    }
-    const TDataSink& output_sink() const;
-
-private:
-    static const std::vector<TScanRangeParams> _no_scan_ranges;
-    static const PerDriverScanRangesMap _no_scan_ranges_per_driver_seq;
-    const TExecPlanFragmentParams& _common_request;
-    const TExecPlanFragmentParams& _unique_request;
-};
-
-class FragmentExecutor {
-public:
-    FragmentExecutor();
-
-    Status prepare(ExecEnv* exec_env, const TExecPlanFragmentParams& common_request,
-                   const TExecPlanFragmentParams& unique_request);
-    Status execute(ExecEnv* exec_env);
-
-    static Status append_incremental_scan_ranges(ExecEnv* exec_env, const TExecPlanFragmentParams& request,
-                                                 TExecPlanFragmentResult* response);
-
-    // Register the partition_value of each scan range into its HiveTableDescriptor's
-    // _partition_id_to_desc_map. The HdfsPartitionDescriptor is allocated from the
-    // query-level ObjectPool (RuntimeStateHelper::global_obj_pool) so that the map's
-    // entries outlive any single fragment instance — see the function body for the
-    // UAF this prevents.
-    //
-    // Exposed here so unit tests can pin the contract; production callers go through
-    // _prepare_exec_plan / append_incremental_scan_ranges, not this entry point.
-    static Status add_scan_ranges_partition_values(RuntimeState* runtime_state,
-                                                   const std::vector<TScanRangeParams>& scan_ranges);
-
-    Status prepare_global_state(ExecEnv* exec_env, const TExecPlanFragmentParams& common_request);
-    void _fail_cleanup(bool fragment_has_registed);
-
-private:
-    uint32_t _calc_dop(ExecEnv* exec_env, const UnifiedExecPlanFragmentParams& request) const;
-    uint32_t _calc_sink_dop(ExecEnv* exec_env, const UnifiedExecPlanFragmentParams& request) const;
-    int _calc_delivery_expired_seconds(const UnifiedExecPlanFragmentParams& request) const;
-    int _calc_query_expired_seconds(const UnifiedExecPlanFragmentParams& request) const;
-
-    // Several steps of prepare a fragment
-    // 1. query context
-    // 2. fragment context
-    // 3. workgroup
-    // 4. runtime state
-    // 5. exec plan
-    // 6. pipeline driver
-    Status _prepare_query_ctx(ExecEnv* exec_env, const UnifiedExecPlanFragmentParams& request);
-    Status _prepare_fragment_ctx(const UnifiedExecPlanFragmentParams& request);
-    Status _prepare_workgroup(const UnifiedExecPlanFragmentParams& request);
-    Status _prepare_runtime_state(ExecEnv* exec_env, const UnifiedExecPlanFragmentParams& request);
-    Status _prepare_exec_plan(ExecEnv* exec_env, const UnifiedExecPlanFragmentParams& request);
-    Status _prepare_global_dict(const UnifiedExecPlanFragmentParams& request);
-    Status _prepare_pipeline_driver(ExecEnv* exec_env, const UnifiedExecPlanFragmentParams& request);
-    Status _prepare_stream_load_pipe(ExecEnv* exec_env, const UnifiedExecPlanFragmentParams& request);
-
-    std::unordered_map<int32_t, ExecutionGroupPtr> _colocate_exec_groups;
-    bool _is_in_colocate_exec_group(PlanNodeId plan_node_id);
-
-    int64_t _fragment_start_time = 0;
-    QueryContextManager* _query_ctx_mgr = nullptr;
-    QueryContext* _query_ctx = nullptr;
-    // Pin the QueryContext alive for at least as long as `_fragment_ctx`.
-    // Fragment operators cache raw pointers into query-scoped state (e.g.
-    // QueryContext::spill_manager()); if the QueryContext is reclaimed (its active
-    // fragment count reaches 0) while this executor still holds the last reference
-    // to `_fragment_ctx`, tearing the fragment down would deref freed memory.
-    // Declared before `_fragment_ctx` so it is destroyed AFTER it.
-    QueryContextPtr _query_ctx_hold = nullptr;
-    FragmentContextPtr _fragment_ctx = nullptr;
-    workgroup::WorkGroupPtr _wg = nullptr;
-};
-} // namespace pipeline
-} // namespace starrocks
