+++ build-support/exec_env_singleton_allowlist.txt
@@ -2,7 +2,7 @@
 src/agent/heartbeat_server.cpp
 src/agent/resource_group_usage_recorder.cpp
 src/agent/task_worker_pool.cpp
-src/exec/pipeline/fragment_executor.cpp
+src/query_orchestration/fragment_executor.cpp
 src/exec/pipeline/pipeline_driver_poller.cpp
 src/service/service_be/config_update_hooks.cpp
 src/runtime/exec_env.cpp
